requirements.txt
Django==3.2.5
psycopg2-binary==2.9.7
djangorestframework==3.12.4
pytest==6.2.5
pytest-django==4.4.0
django-allauth==0.54.0
stripe==5.5.0
requests==2.31.0
requests-oauthlib==1.3.1